refactor: extract strip_fences into shared hermia/normalize.py
Single source of truth for fence-stripping, shared by the grader and the
upcoming reproducibility canonical-match. Behavior-preserving move; updates
all 3 production importers (runner, corpus_audit/mining, corpus_audit/confusion)
and relocates the 6 tests to test_normalize.py.
